requirements.txt
pandas
numpy
scipy
scikit-learn
biopython
gensim==3.8.3
jsonlines
scann
h5py